Refactor: API 모듈화, pages 모두 완료.
