+++ src/components/investmentStatus/InvestmentStatus.jsx
@@ -11,7 +11,6 @@ import { SimpleButton } from "../buttons/Buttons";
 import { getInvestmentByCompanyId } from "../../api/investment.api";
 import table from "../../styles/table.module.scss";
 import { SimpleButton } from "../buttons/Buttons";
-import { getInvestmentByCompanyId } from "../../api/investment.api";
 
 const InvestmentStatus = () => {
   const { id } = useParams();
