--- src/env.js
@@ -1,2 +1,0 @@
-export const dataUrl =
-  "https://port-0-viewmystartup-3-m8ml2ohm3e1c28b1.sel4.cloudtype.app";
